Rearranging the last of the macros (attack, hotbar, template)
