+++ src/module/macros/item.ts
@@ -12,15 +12,13 @@ import {
 } from "machine-mind";
 import { applyCollapseListeners } from "../helpers/collapse";
 import { getMacroSpeaker, ownedItemFromString } from "./util"
+import { prepareAttackMacro } from "./attack"
 import { rollReactionMacro } from "./reaction"
 import { rollSystemMacro } from "./system"
 import { rollTalentMacro } from "./talent"
 import { prepareTechMacro } from "./tech"
 import { rollTextMacro } from "./text"
 import { rollTriggerMacro } from "./trigger"
-import {
-  prepareAttackMacro,
-} from "../macros"
 
 const lp = LANCER.log_prefix;
